posting list v2 + bug fixed

+++ src/main/java/org/common/DocInfo.java
@@ -17,7 +17,7 @@ public class DocInfo {
         return pid;
     }
 
-    public long getLength() {
+    public int getLength() {
         return length;
     }
